+++ internal/codec/reader.go
@@ -40,6 +40,7 @@ var ReaderDocs = docs.FieldString(
 	"lines", "Consume the file in segments divided by linebreaks.",
 	"multipart", "Consumes the output of another codec and batches messages together. A batch ends when an empty message is consumed. For example, the codec `lines/multipart` could be used to consume multipart messages where an empty line indicates the end of each batch.",
 	"regex:(?m)^\\d\\d:\\d\\d:\\d\\d", "Consume the file in segments divided by regular expression.",
+	"skipbom", "Skip one or more byte order marks for each opened reader, this codec should precede another codec, e.g. `skipbom/csv`, etc.",
 	"tar", "Parse the file as a tar archive, and consume each file of the archive as a message.",
 )
 
@@ -193,7 +194,14 @@ func ioReader(codec string, conf ReaderConfig) (ioReaderConstructor, bool) {
 				r.Close()
 				return nil, err
 			}
-			return g, nil
+			unzipped := ioReadCloserWrapper{Reader: g, underlying: r}
+			return &unzipped, nil
+		}, true
+	}
+	if codec == "skipbom" {
+		return func(_ string, r io.ReadCloser) (io.ReadCloser, error) {
+			skipBom := ioReadCloserWrapper{Reader: skipBOM(r), underlying: r}
+			return &skipBom, nil
 		}, true
 	}
 	return nil, false
@@ -340,6 +348,21 @@ func autoCodec(conf ReaderConfig) ReaderConstructor {
 		}
 		return ctor(path, r, fn)
 	}
+}
+
+// ioReadCloserWrapper is a helper that closes both the upper and underlying reader
+// when you are creating some sort of wrapped reader where you want to ensure both
+// are closed.
+type ioReadCloserWrapper struct {
+	io.Reader
+	underlying io.ReadCloser
+}
+
+func (w ioReadCloserWrapper) Close() error {
+	if rc, ok := w.Reader.(io.Closer); ok {
+		rc.Close()
+	}
+	return w.underlying.Close()
 }
 
 //------------------------------------------------------------------------------